Abstract construction integration tests
